requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pyodbc==5.2.0
azure-identity==1.19.0
azure-monitor-opentelemetry==1.6.4
pydantic==2.10.6
